pyproject.toml
[project]
name = "ai-art"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.12",
    "openai>=1.64.0",
    "python-dotenv>=1.0.1",
    "ruff>=0.9.7",
    "yagmail>=0.15.293",
]
